Changes as per the review
